Moved mount package outside sys package and deprecated old methods
