+++ src/main/java/io/github/jopenlibs/vault/api/sys/mounts/MountType.java
@@ -1,4 +1,4 @@
-package io.github.jopenlibs.vault.api.mounts;
+package io.github.jopenlibs.vault.api.sys.mounts;
 
 import java.util.Arrays;
