+++ src/main/java/io/github/jopenlibs/vault/api/sys/mounts/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * <p>Classes implementing Vault's mounts system.</p>
+ *
+ * <p>The classes in this package are not meant to be instantiated directly.  Rather, they should
+ * be used by way of the <code>io.github.jopenlibs.vault.Vault</code> driver class, in a DSL-style
+ * builder pattern approach.</p>
+ */
+package io.github.jopenlibs.vault.api.sys.mounts;
